backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
google-generativeai==0.4.0
python-dotenv==1.0.0
slowapi==0.1.9
cachetools==5.3.2
pydantic==2.5.3
python-multipart==0.0.6
